Develop

+++ Plugins/UnLua/Source/UnLua/UnLua.Build.cs
@@ -12,134 +12,84 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
 // See the License for the specific language governing permissions and limitations under the License.
 
+using System;
 using System.IO;
+#if UE_5_0_OR_LATER
+using EpicGames.Core;
+#else
+using Tools.DotNETCommon;
+#endif
 using UnrealBuildTool;
 
 public class UnLua : ModuleRules
 {
-	public UnLua(ReadOnlyTargetRules Target) : base(Target)
+    public UnLua(ReadOnlyTargetRules Target) : base(Target)
     {
         SetupScripts();
 
         bEnforceIWYU = false;
-
-        PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
 
         PublicIncludePaths.AddRange(
-			new string[] {
-            }
-			);
-
-
-		PrivateIncludePaths.AddRange(
-			new string[] {
-                "UnLua/Private",
-            }
-            );
-
-
-        PublicIncludePathModuleNames.AddRange(
-            new string[] {
-                "ApplicationCore",
+            new string[]
+            {
             }
         );
 
+        PrivateIncludePaths.AddRange(
+            new[]
+            {
+                "UnLua/Private",
+            }
+        );
 
-        PrivateDependencyModuleNames.AddRange(
-			new string[]
-			{
+        PublicDependencyModuleNames.AddRange(
+            new[]
+            {
                 "Core",
                 "CoreUObject",
-				"Engine",
+                "Engine",
                 "Slate",
                 "InputCore",
-                "Projects",
                 "Lua"
-			}
-			);
-		
+            }
+        );
+
         PublicIncludePaths.Add(Path.Combine(ModuleDirectory, "Private"));
-		
-        if (Target.bBuildEditor == true)
-        {
+
+        if (Target.bBuildEditor)
             PrivateDependencyModuleNames.Add("UnrealEd");
-        }
 
-        bool bAutoStartup = true;
-        if (bAutoStartup)
-        {
-            PublicDefinitions.Add("AUTO_UNLUA_STARTUP=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("AUTO_UNLUA_STARTUP=0");
-        }
+        var projectDir = Target.ProjectFile.Directory;
+        var configFilePath = projectDir + "/Config/DefaultUnLuaEditor.ini";
+        var configFileReference = new FileReference(configFilePath); 
+        var configFile = FileReference.Exists(configFileReference) ? new ConfigFile(configFileReference) : new ConfigFile();
+        var config = new ConfigHierarchy(new[] { configFile });
+        const string Section = "/Script/UnLuaEditor.UnLuaEditorSettings";
 
-        bool bWithUE4Namespace = true;
-        if (bWithUE4Namespace)
+        Action<string, string, bool> loadBoolConfig = (key, macro, defaultValue) =>
         {
-            PublicDefinitions.Add("WITH_UE4_NAMESPACE=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("WITH_UE4_NAMESPACE=0");
-        }
-    
-        bool bSupportsRpcCall = true;
-        if (bSupportsRpcCall)
-        {
-            PublicDefinitions.Add("SUPPORTS_RPC_CALL=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("SUPPORTS_RPC_CALL=0");
-        }
+            bool flag;
+            if (!config.GetBool(Section, key, out flag))
+                flag = defaultValue;
+            PublicDefinitions.Add(string.Format("{0}={1}", macro, (flag ? "1" : "0")));
+        };
 
-        bool bSupportsCommandlet = true;
-        if (bSupportsCommandlet)
-        {
-            PublicDefinitions.Add("SUPPORTS_COMMANDLET=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("SUPPORTS_COMMANDLET=0");
-        }
-
-        bool bEnableAutoCleanNoneNativeClass = true;
-        if (bEnableAutoCleanNoneNativeClass)
-        {
-            PublicDefinitions.Add("ENABLE_AUTO_CLEAN_NNATCLASS=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("ENABLE_AUTO_CLEAN_NNATCLASS=0");
-        }
-
-        bool bEnableTypeCheck = true;
-        if (bEnableTypeCheck)
-        {
-            PublicDefinitions.Add("ENABLE_TYPE_CHECK=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("ENABLE_TYPE_CHECK=0");
-        }
-
-        bool bEnableDebug = false;
-        if (bEnableDebug)
-        {
-            PublicDefinitions.Add("UNLUA_ENABLE_DEBUG=1");
-        }
-        else
-        {
-            PublicDefinitions.Add("UNLUA_ENABLE_DEBUG=0");
-        }
-
+        loadBoolConfig("bAutoStartup", "AUTO_UNLUA_STARTUP", true);
+        loadBoolConfig("bEnableDebug", "UNLUA_ENABLE_DEBUG", false);
+        loadBoolConfig("bEnablePersistentParamBuffer", "ENABLE_PERSISTENT_PARAM_BUFFER", true);
+        loadBoolConfig("bEnableTypeChecking", "ENABLE_TYPE_CHECK", true);
+        loadBoolConfig("bEnableRPCCall", "SUPPORTS_RPC_CALL", true);
+        loadBoolConfig("bEnableCallOverriddenFunction", "ENABLE_CALL_OVERRIDDEN_FUNCTION", true);
+        loadBoolConfig("bLuaCompileAsCpp", "LUA_COMPILE_AS_CPP", false);
+        loadBoolConfig("bWithUE4Namespace", "WITH_UE4_NAMESPACE", true);
+        loadBoolConfig("bLegacyReturnOrder", "UNLUA_LEGACY_RETURN_ORDER", false);
+        loadBoolConfig("bLegacyBlueprintPath", "UNLUA_LEGACY_BLUEPRINT_PATH", false);
     }
 
     private void SetupScripts()
     {
-        const string UnLuaSourceFileName = "UnLua.lua";
+        const string UnLuaSourceFileName = "UnLuaHotReload.lua";
         var PluginContentDirectory = Path.Combine(PluginDirectory, "Content");
         var DefaultScriptDirectory = Path.Combine(Target.ProjectFile.Directory.ToString(), "Content/Script");
         if (!Directory.Exists(DefaultScriptDirectory))
@@ -147,7 +97,7 @@ public class UnLua : ModuleRules
 
         var SrcPath = Path.Combine(PluginContentDirectory, UnLuaSourceFileName);
         var DstPath = Path.Combine(DefaultScriptDirectory, UnLuaSourceFileName);
-        if (!File.Exists(DstPath))
+        if (!File.Exists(DstPath) && File.Exists(SrcPath))
             File.Copy(SrcPath, DstPath);
     }
 }
